pdf_extractor/requirements/base.txt
# Base requirements — app MUST run with only these installed
click>=8.1.0
rich>=13.0.0
flask>=3.0.0
pymupdf>=1.23.0
pdfplumber>=0.9.0
pymupdf4llm>=0.0.5
pypdf>=3.0.0
pdfminer.six>=20221105
tabulate>=0.9.0
tqdm>=4.65.0
langdetect>=1.0.9
chardet>=5.0.0
Pillow>=10.0.0
python-docx>=1.1.0
python-pptx>=0.6.21
